requirements.txt
fastparquet
irrcac==0.4.0
matplotlib
numpy
pandas
scikit-learn
scip
seaborn
tqdm
typing_extensions